Walk: enforce `MAX_STAT` during ScanDirectory()
As a quick kludge, this commit creates a copy of ScanDirectory() as a
coroutine so we can `co_await resume_stat` (we stillneed the
synchronous version for method Start()).  The coroutine copy can then
enforce `MAX_STAT` to limit the io_uring statx() concurrency; this is
necessary because cache directories with millions of cached files
would cause the kernel to consume gigabytes of memory and exceed our
memory limit.