cull 2% more in each run (to compensate for new files being added)
If each Cull operation attempts to reach exactly BRUN / FRUN, it will
fail to meet that goal because new files are being added during the
Cull, and another Cull will be started right after that, which will
again fail to meet the goal, leaving the daemon in an endless culling
loop.

Adding 2% to the configured BRUN / FRUN values is an experiment to see
whether this solves the loop.